move executesqlfile to library
