+++ gpx2db_pkg/gpx2db/__init__.py
@@ -1,1 +1,2 @@
 from .gpx2dblib import Gpx2db
+from .utils import ExecuteSQLFile
